Refactor: PR 내용 반영
- DeliveryType 이름 DeliveryPriceOption로 변경
- NONE -> FREE로 변경
