+++ BE/src/main/java/com/example/be/controller/dish/dto/DishSaveRequest.java
@@ -1,7 +1,7 @@
 package com.example.be.controller.dish.dto;
 
 import com.example.be.domain.dish.Badge;
-import com.example.be.domain.dish.DeliveryType;
+import com.example.be.domain.dish.DeliveryPriceOption;
 import com.example.be.domain.dish.Dish;
 import com.example.be.domain.dish.DishStatus;
 import lombok.Getter;
@@ -15,25 +15,25 @@ public class DishSaveRequest {
     private BigDecimal normalPrice;
     private BigDecimal salePrice;
     private Badge badge;
-    private DeliveryType deliveryType;
+    private DeliveryPriceOption deliveryPriceOption;
     private String thumbnail;
     private DishStatus dishStatus;
     private int count;
 
-    public DishSaveRequest(String name, String description, BigDecimal normalPrice, BigDecimal salePrice, Badge badge, DeliveryType deliveryType, String thumbnail, DishStatus dishStatus, int count) {
+    public DishSaveRequest(String name, String description, BigDecimal normalPrice, BigDecimal salePrice, Badge badge, DeliveryPriceOption deliveryPriceOption, String thumbnail, DishStatus dishStatus, int count) {
         this.name = name;
         this.description = description;
         this.normalPrice = normalPrice;
         this.salePrice = salePrice;
         this.badge = badge;
-        this.deliveryType = deliveryType;
+        this.deliveryPriceOption = deliveryPriceOption;
         this.thumbnail = thumbnail;
         this.dishStatus = dishStatus;
         this.count = count;
     }
 
     public Dish toEntity() {
-        return new Dish(name, description, normalPrice, badge, deliveryType, thumbnail, dishStatus, count);
+        return new Dish(name, description, normalPrice, badge, deliveryPriceOption, thumbnail, dishStatus, count);
     }
 
     @Getter
