+++ BE/src/main/java/com/example/be/domain/user/Bio.java
@@ -1,0 +1,6 @@
+package com.example.be.domain.user;
+
+public enum Bio {
+    MAN,
+    WOMAN;
+}
